Made SceneManager more efficient
Render data is now only updated for an entity when it recieves an update
event for that entity.
And right before render the render and light data is gathered to be
rendered, instead of every udpate tick.

+++ AlphaEngine/Source/Logic/LogicSystem.cpp
@@ -74,7 +74,7 @@ namespace alpha
             //auto entity_id = key_value.first;
             auto entity = key_value.second;
 
-            if (entity->Update(fCurrentTime, fElapsedTime))
+            if (entity->Update(static_cast<float>(fCurrentTime), static_cast<float>(fElapsedTime)))
             {
                 this->PublishEvent(new Event_EntityUpdated(entity));
             }
